First pass at factoring FogHelper and TopoHelper out into an ec-common cookbook
