+++ cookbooks/ec-common/README.md
@@ -1,0 +1,4 @@
+# ec-common
+
+TODO: Enter the cookbook description here.
+
