+++ cookbooks/ec-common/libraries/topo_helper.rb
