+++ cookbooks/ec-common/metadata.rb
@@ -1,9 +1,8 @@
-name             'ec-harness'
+name             'ec-common'
 maintainer       'Irving Popovetsky'
 maintainer_email 'irving@getchef.com'
 license          'All rights reserved'
-description      'Installs/Configures ec-harness, the Enterprise Chef HA test-harness'
+description      'Installs/Configures ec-common'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends 'ec-common'
