+++ cookbooks/private-chef/metadata.rb
@@ -8,3 +8,4 @@ depends 'hostsfile'
 depends 'lvm'
 depends 'yum'
 depends 'ec-tools'
+depends 'ec-common'
